Added error handler uses cases examples docs
